+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 0.12.0-dev
+Version 0.12.0
 - BC break: renamed IXmlConvertible to XmlConvertible and IRssExtension to RssExtension
 - BC break: replaced Nette events with PSR-14 event dispatcher
 - BC break: Generator::$dataSource now only accepts Closure or null